refactor: unify format parsing
