+++ internal/setup/setup.go
@@ -13,6 +13,7 @@ import (
 	"gopkg.in/yaml.v3"
 
 	"github.com/canonical/chisel/internal/deb"
+	"github.com/canonical/chisel/internal/pgputil"
 	"github.com/canonical/chisel/internal/strdist"
 )
 
@@ -270,6 +271,7 @@ func readRelease(baseDir string) (*Release, error) {
 	}
 	return release, err
 }
+
 func readSlices(release *Release, baseDir, dirName string) error {
 	entries, err := os.ReadDir(dirName)
 	if err != nil {
@@ -316,36 +318,22 @@ func readSlices(release *Release, baseDir, dirName string) error {
 	return nil
 }
 
-var ubuntuAdjectives = map[string]string{
-	"18.04": "bionic",
-	"20.04": "focal",
-	"22.04": "jammy",
-	"22.10": "kinetic",
+type yamlRelease struct {
+	Format   string                 `yaml:"format"`
+	Archives map[string]yamlArchive `yaml:"archives"`
+	PubKeys  map[string]yamlPubKey  `yaml:"public-keys"`
+	// V1PubKeys is used for compatibility with format "chisel-v1".
+	V1PubKeys map[string]yamlPubKey `yaml:"v1-public-keys"`
 }
 
-func parseRelease(baseDir, filePath string, data []byte) (*Release, error) {
-	type yamlFormat struct {
-		Format string `yaml:"format"`
-	}
-
-	fileName := stripBase(baseDir, filePath)
-
-	yamlVar := yamlFormat{}
-	dec := yaml.NewDecoder(bytes.NewBuffer(data))
-	dec.KnownFields(false)
-	err := dec.Decode(&yamlVar)
-	if err != nil {
-		return nil, fmt.Errorf("%s: cannot parse release definition: %v", fileName, err)
-	}
-
-	switch yamlVar.Format {
-	case "chisel-v1":
-		return parseReleaseChiselV1(baseDir, filePath, data)
-	case "v1":
-		return parseReleaseV1(baseDir, filePath, data)
-	default:
-		return nil, fmt.Errorf("%s: unknown format %q", fileName, yamlVar.Format)
-	}
+type yamlArchive struct {
+	Version    string   `yaml:"version"`
+	Suites     []string `yaml:"suites"`
+	Components []string `yaml:"components"`
+	Default    bool     `yaml:"default"`
+	PubKeys    []string `yaml:"public-keys"`
+	// V1PubKeys is used for compatibility with format "chisel-v1".
+	V1PubKeys []string `yaml:"v1-public-keys"`
 }
 
 type yamlPackage struct {
@@ -401,6 +389,112 @@ type yamlSlice struct {
 	Essential []string             `yaml:"essential"`
 	Contents  map[string]*yamlPath `yaml:"contents"`
 	Mutate    string               `yaml:"mutate"`
+}
+
+type yamlPubKey struct {
+	ID    string `yaml:"id"`
+	Armor string `yaml:"armor"`
+}
+
+var ubuntuAdjectives = map[string]string{
+	"18.04": "bionic",
+	"20.04": "focal",
+	"22.04": "jammy",
+	"22.10": "kinetic",
+}
+
+func parseRelease(baseDir, filePath string, data []byte) (*Release, error) {
+	release := &Release{
+		Path:     baseDir,
+		Packages: make(map[string]*Package),
+		Archives: make(map[string]*Archive),
+	}
+
+	fileName := stripBase(baseDir, filePath)
+
+	yamlVar := yamlRelease{}
+	dec := yaml.NewDecoder(bytes.NewBuffer(data))
+	dec.KnownFields(false)
+	err := dec.Decode(&yamlVar)
+	if err != nil {
+		return nil, fmt.Errorf("%s: cannot parse release definition: %v", fileName, err)
+	}
+	if yamlVar.Format != "chisel-v1" && yamlVar.Format != "v1" {
+		return nil, fmt.Errorf("%s: unknown format %q", fileName, yamlVar.Format)
+	}
+	// If format is "chisel-v1" we have to translate from the yaml key "v1-public-keys" to
+	// "public-keys".
+	if yamlVar.Format == "chisel-v1" {
+		yamlVar.PubKeys = yamlVar.V1PubKeys
+		for name, details := range yamlVar.Archives {
+			details.PubKeys = details.V1PubKeys
+			yamlVar.Archives[name] = details
+		}
+	}
+	if len(yamlVar.Archives) == 0 {
+		return nil, fmt.Errorf("%s: no archives defined", fileName)
+	}
+
+	// Decode the public keys and match against provided IDs.
+	pubKeys := make(map[string]*packet.PublicKey, len(yamlVar.PubKeys))
+	for keyName, yamlPubKey := range yamlVar.PubKeys {
+		key, err := pgputil.DecodePubKey([]byte(yamlPubKey.Armor))
+		if err != nil {
+			return nil, fmt.Errorf("%s: cannot decode public key %q: %w", fileName, keyName, err)
+		}
+		if yamlPubKey.ID != key.KeyIdString() {
+			return nil, fmt.Errorf("%s: public key %q armor has incorrect ID: expected %q, got %q", fileName, keyName, yamlPubKey.ID, key.KeyIdString())
+		}
+		pubKeys[keyName] = key
+	}
+
+	for archiveName, details := range yamlVar.Archives {
+		if details.Version == "" {
+			return nil, fmt.Errorf("%s: archive %q missing version field", fileName, archiveName)
+		}
+		if len(details.Suites) == 0 {
+			adjective := ubuntuAdjectives[details.Version]
+			if adjective == "" {
+				return nil, fmt.Errorf("%s: archive %q missing suites field", fileName, archiveName)
+			}
+			details.Suites = []string{adjective}
+		}
+		if len(details.Components) == 0 {
+			return nil, fmt.Errorf("%s: archive %q missing components field", fileName, archiveName)
+		}
+		if len(yamlVar.Archives) == 1 {
+			details.Default = true
+		} else if details.Default && release.DefaultArchive != "" {
+			return nil, fmt.Errorf("%s: more than one default archive: %s, %s", fileName, release.DefaultArchive, archiveName)
+		}
+		if details.Default {
+			release.DefaultArchive = archiveName
+		}
+		if len(details.PubKeys) == 0 {
+			if yamlVar.Format == "chisel-v1" {
+				return nil, fmt.Errorf("%s: archive %q missing v1-public-keys field", fileName, archiveName)
+			} else {
+				return nil, fmt.Errorf("%s: archive %q missing public-keys field", fileName, archiveName)
+			}
+		}
+		var archiveKeys []*packet.PublicKey
+		for _, keyName := range details.PubKeys {
+			key, ok := pubKeys[keyName]
+			if !ok {
+				return nil, fmt.Errorf("%s: archive %q refers to undefined public key %q", fileName, archiveName, keyName)
+			}
+			archiveKeys = append(archiveKeys, key)
+		}
+		release.Archives[archiveName] = &Archive{
+			Name:       archiveName,
+			Version:    details.Version,
+			Suites:     details.Suites,
+			Components: details.Components,
+			PubKeys:    archiveKeys,
+		}
+	}
+
+	return release, err
 }
 
 func parsePackage(baseDir, pkgName, pkgPath string, data []byte) (*Package, error) {
